feat: restructure
