+++ packages/archive-reader/src/cover/resolveArchiveCover.ts
@@ -1,8 +1,8 @@
 import { detectMimeTypeFromName, parseContentType } from "@prose-reader/shared"
 import type { Archive } from "../archives/types.ts"
 import { getArchiveFileRecordByUri } from "../archives/types.ts"
-import type { ArchiveOpfParsed } from "../opf/readArchiveOpf.ts"
-import { readArchiveOpf } from "../opf/readArchiveOpf.ts"
+import type { ArchiveOpfParsed } from "../metadata/opf/readArchiveOpf.ts"
+import { readArchiveOpf } from "../metadata/opf/readArchiveOpf.ts"
 import type { ArchiveReadingOrderItem } from "../readingOrder/resolveArchiveReadingOrder.ts"
 import { resolveArchiveReadingOrder } from "../readingOrder/resolveArchiveReadingOrder.ts"
 import type { ResolvedCover } from "../types/resolvedMetadata.ts"
