+++ packages/archive-reader/src/metadata/apple/parse.test.ts
