+++ packages/archive-reader/src/metadata/apple/parse.ts
