+++ packages/archive-reader/src/metadata/apple/readArchiveApple.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from "vitest"
-import { createArchive } from "../archives/createArchive"
-import { blobFileAccessors } from "../archives/fileAccessors"
+import { createArchive } from "../../archives/createArchive"
+import { blobFileAccessors } from "../../archives/fileAccessors"
 import { readArchiveApple } from "./readArchiveApple"
 
 const textRecord = (uri: string, content = ``) => ({
