+++ packages/archive-reader/src/metadata/apple/resolve.test.ts
