+++ packages/archive-reader/src/metadata/comicInfo/manga.test.ts
