+++ packages/archive-reader/src/metadata/comicInfo/manga.ts
