+++ packages/archive-reader/src/metadata/comicInfo/parse.ts
