+++ packages/archive-reader/src/metadata/opf/spineItemrefProperties.test.ts
