+++ packages/archive-reader/src/resolve.ts
@@ -1,11 +1,11 @@
-import type { AppleMetadata } from "./apple/parse.ts"
-import { resolveApple } from "./apple/resolve.ts"
-import type { ComicInfo } from "./comicInfo/parse.ts"
-import { resolveComicInfo } from "./comicInfo/resolve.ts"
-import type { KoboMetadata } from "./kobo/parse.ts"
-import { resolveKobo } from "./kobo/resolve.ts"
-import type { OpfMetadata } from "./opf/parse.ts"
-import { resolveOpf } from "./opf/resolve.ts"
+import type { AppleMetadata } from "./metadata/apple/parse.ts"
+import { resolveApple } from "./metadata/apple/resolve.ts"
+import type { ComicInfo } from "./metadata/comicInfo/parse.ts"
+import { resolveComicInfo } from "./metadata/comicInfo/resolve.ts"
+import type { KoboMetadata } from "./metadata/kobo/parse.ts"
+import { resolveKobo } from "./metadata/kobo/resolve.ts"
+import type { OpfMetadata } from "./metadata/opf/parse.ts"
+import { resolveOpf } from "./metadata/opf/resolve.ts"
 import type { ResolvedMetadata } from "./types/resolvedMetadata.ts"
 
 export type ResolvedArchiveInput =
