+++ packages/archive-reader/src/toc/ncx.ts
@@ -2,7 +2,7 @@ import { getUriBasePath, urlJoin } from "@prose-reader/shared"
 import { XmlDocument, type XmlElement } from "xmldoc"
 import { readRecordAsText } from "../archives/readRecordAsText.ts"
 import type { Archive } from "../archives/types.ts"
-import type { OpfMetadata } from "../opf/parse.ts"
+import type { OpfMetadata } from "../metadata/opf/parse.ts"
 import type { ArchiveTocItem } from "./types.ts"
 
 const mapNcxChapter = (
