+++ packages/archive-reader/src/toc/resolveArchiveToc.test.ts
@@ -3,7 +3,7 @@ import path from "node:path"
 import { describe, expect, it } from "vitest"
 import { createArchive } from "../archives/createArchive"
 import { blobFileAccessors } from "../archives/fileAccessors"
-import { readArchiveOpf } from "../opf/readArchiveOpf"
+import { readArchiveOpf } from "../metadata/opf/readArchiveOpf"
 import { resolveArchiveToc } from "./resolveArchiveToc"
 
 const textRecord = (uri: string, content = ``) => ({
